# server/openapi.yaml
openapi: 3.1.0
info:
  title: Finance Query v2
  version: 3.0.0
  description: High-performance financial data API server built on finance-query library
servers:
  - url: "https://finance-query.com"
  - url: "http://localhost:8000"
paths:
...
  "/v2/industries/{industry}":
    get:
      operationId: getIndustry
      summary: Get industry data
      parameters:
        - name: fields
          in: query
          required: false
          schema:
            description: Comma-separated list of fields to include in response
            type: string
        - name: format
          in: query
          required: false
          schema:
            description: "Value format: raw, pretty, or both (default: raw)"
            $ref: "#/components/schemas/ValueFormat"
        - name: lang
          in: query
          required: false
          schema:
            description: "Target language for translated text fields (BCP 47, e.g. \"ja\", \"zh-Hant\");\nfalls back to the Accept-Language header"
            type: string
        - name: industry
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Industry"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GqlIndustryData"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Symbol or resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate limited (global token bucket)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
          headers:
            Retry-After:
              schema:
                type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  "/v2/sectors/{sector}":
    get:
      operationId: getSector
      summary: Get sector data
      parameters:
        - name: fields
          in: query
          required: false
          schema:
            description: Comma-separated list of fields to include in response
            type: string
        - name: format
          in: query
          required: false
          schema:
            description: "Value format: raw, pretty, or both (default: raw)"
            $ref: "#/components/schemas/ValueFormat"
        - name: lang
          in: query
          required: false
          schema:
            description: "Target language for translated text fields (BCP 47, e.g. \"ja\", \"zh-Hant\");\nfalls back to the Accept-Language header"
            type: string
        - name: sector
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Sector"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GqlSectorData"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Symbol or resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate limited (global token bucket)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
          headers:
            Retry-After:
              schema:
                type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      description: "Generic REST error body, emitted on 4xx/5xx by the GraphQL bridge and by\nhandlers rejecting invalid path/query input."
      type: object
      properties:
        error:
          description: Human-readable error message.
          type: string
        status:
          description: "HTTP status code, duplicated in the body."
          type: integer
          format: int32
          minimum: 0
      required:
        - error
        - status
...
    GqlBenchmarkPerformance:
      type: object
      properties:
        dayChangePercent:
          type: number
          format: double
        fiveYearChangePercent:
          type: number
          format: double
        name:
          type: string
        oneYearChangePercent:
          type: number
          format: double
        threeYearChangePercent:
          type: number
          format: double
        ytdChangePercent:
          type: number
          format: double
...
    GqlFormattedObject:
      description: "The `{raw, fmt, longFmt}` wrapper as it reaches the wire under `format=both`."
      type: object
      properties:
        fmt:
          description: "Human-readable rendering, e.g. `\"182.50\"` or `\"3.50T\"`."
          type: string
        longFmt:
          description: "Fully expanded rendering for large numbers, e.g. `\"3,500,000,000,000\"`."
          type: string
        raw:
          description: "The underlying datum. May be `null` when upstream reports no value."
      required:
        - raw
      additionalProperties: false
    GqlFormattedValue:
      description: "One Yahoo Finance formatted value, in whichever shape the request's `format` selected."
      oneOf:
        - type: integer
          format: int64
        - type: number
          format: double
        - type: string
        - $ref: "#/components/schemas/GqlFormattedObject"
        - {}
...
    GqlGrowthCompany:
      type: object
      properties:
        growthEstimate:
          type: number
          format: double
        lastPrice:
          type: number
          format: double
        name:
          type: string
        symbol:
          type: string
        ytdReturn:
          type: number
          format: double
      required:
        - symbol
...
    GqlIndustryCompany:
      type: object
      properties:
        dayChangePercent:
          type: number
          format: double
        lastPrice:
          type: number
          format: double
        marketCap:
          type: number
          format: double
        marketWeight:
          type: number
          format: double
        name:
          type: string
        rating:
          type: string
        symbol:
          type: string
        targetPrice:
          type: number
          format: double
        ytdReturn:
          type: number
          format: double
      required:
        - symbol
    GqlIndustryData:
      type: object
      properties:
        benchmark:
          $ref: "#/components/schemas/GqlBenchmarkPerformance"
        key:
          type: string
        name:
          type: string
        overview:
          $ref: "#/components/schemas/GqlIndustryOverview"
        performance:
          $ref: "#/components/schemas/GqlIndustryPerformance"
        researchReports:
          type: array
          items:
            $ref: "#/components/schemas/GqlIndustryResearchReport"
        sectorKey:
          type: string
        sectorName:
          type: string
        symbol:
          type: string
        topCompanies:
          type: array
          items:
            $ref: "#/components/schemas/GqlIndustryCompany"
        topGrowthCompanies:
          type: array
          items:
            $ref: "#/components/schemas/GqlGrowthCompany"
        topPerformingCompanies:
          type: array
          items:
            $ref: "#/components/schemas/GqlPerformingCompany"
      required:
        - name
        - key
        - topCompanies
        - topPerformingCompanies
        - topGrowthCompanies
        - researchReports
    GqlIndustryOverview:
      type: object
      properties:
        companiesCount:
          type: integer
          format: int64
        description:
          type: string
        employeeCount:
          type: integer
          format: int64
        marketCap:
          type: number
          format: double
        marketWeight:
          type: number
          format: double
    GqlIndustryPerformance:
      type: object
      properties:
        dayChangePercent:
          type: number
          format: double
        fiveYearChangePercent:
          type: number
          format: double
        oneYearChangePercent:
          type: number
          format: double
        threeYearChangePercent:
          type: number
          format: double
        ytdChangePercent:
          type: number
          format: double
    GqlIndustryResearchReport:
      type: object
      properties:
        id:
          type: string
        investmentRating:
          type: string
        provider:
          type: string
        reportDate:
          type: string
        reportType:
          type: string
        targetPrice:
          type: number
          format: double
        targetPriceStatus:
          type: string
        title:
          type: string
...
    GqlPerformingCompany:
      type: object
      properties:
        lastPrice:
          type: number
          format: double
        name:
          type: string
        symbol:
          type: string
        targetPrice:
          type: number
          format: double
        ytdReturn:
          type: number
          format: double
      required:
        - symbol
...
    GqlSectorCompany:
      type: object
      properties:
        dayChangePercent:
          $ref: "#/components/schemas/GqlFormattedValue"
        lastPrice:
          $ref: "#/components/schemas/GqlFormattedValue"
        marketCap:
          $ref: "#/components/schemas/GqlFormattedValue"
        marketWeight:
          $ref: "#/components/schemas/GqlFormattedValue"
        name:
          type: string
        rating:
          type: string
        symbol:
          type: string
        targetPrice:
          $ref: "#/components/schemas/GqlFormattedValue"
        ytdReturn:
          $ref: "#/components/schemas/GqlFormattedValue"
      required:
        - symbol
    GqlSectorData:
      type: object
      properties:
        benchmark:
          $ref: "#/components/schemas/GqlSectorPerformance"
        benchmarkName:
          type: string
        industries:
          type: array
          items:
            $ref: "#/components/schemas/GqlSectorIndustry"
        key:
          type: string
        name:
          type: string
        overview:
          $ref: "#/components/schemas/GqlSectorOverview"
        performance:
          $ref: "#/components/schemas/GqlSectorPerformance"
        researchReports:
          type: array
          items:
            $ref: "#/components/schemas/GqlSectorResearchReport"
        symbol:
          type: string
        topCompanies:
          type: array
          items:
            $ref: "#/components/schemas/GqlSectorCompany"
        topEtfs:
          type: array
          items:
            $ref: "#/components/schemas/GqlSectorETF"
        topMutualFunds:
          type: array
          items:
            $ref: "#/components/schemas/GqlSectorMutualFund"
      required:
        - name
        - key
        - topCompanies
        - topEtfs
        - topMutualFunds
        - industries
        - researchReports
    GqlSectorETF:
      type: object
      properties:
        expenseRatio:
          $ref: "#/components/schemas/GqlFormattedValue"
        lastPrice:
          $ref: "#/components/schemas/GqlFormattedValue"
        name:
          type: string
        netAssets:
          $ref: "#/components/schemas/GqlFormattedValue"
        symbol:
          type: string
        ytdReturn:
          $ref: "#/components/schemas/GqlFormattedValue"
      required:
        - symbol
    GqlSectorIndustry:
      type: object
      properties:
        dayChangePercent:
          $ref: "#/components/schemas/GqlFormattedValue"
        key:
          type: string
        marketWeight:
          $ref: "#/components/schemas/GqlFormattedValue"
        name:
          type: string
        symbol:
          type: string
        ytdReturn:
          $ref: "#/components/schemas/GqlFormattedValue"
      required:
        - name
    GqlSectorMutualFund:
      type: object
      properties:
        expenseRatio:
          $ref: "#/components/schemas/GqlFormattedValue"
        lastPrice:
          $ref: "#/components/schemas/GqlFormattedValue"
        name:
          type: string
        netAssets:
          $ref: "#/components/schemas/GqlFormattedValue"
        symbol:
          type: string
        ytdReturn:
          $ref: "#/components/schemas/GqlFormattedValue"
      required:
        - symbol
    GqlSectorOverview:
      type: object
      properties:
        companiesCount:
          type: integer
          format: int64
        description:
          type: string
        employeeCount:
          $ref: "#/components/schemas/GqlFormattedValue"
        industriesCount:
          type: integer
          format: int64
        marketCap:
          $ref: "#/components/schemas/GqlFormattedValue"
        marketWeight:
          $ref: "#/components/schemas/GqlFormattedValue"
    GqlSectorPerformance:
      type: object
      properties:
        dayChangePercent:
          $ref: "#/components/schemas/GqlFormattedValue"
        fiveYearChangePercent:
          $ref: "#/components/schemas/GqlFormattedValue"
        oneYearChangePercent:
          $ref: "#/components/schemas/GqlFormattedValue"
        threeYearChangePercent:
          $ref: "#/components/schemas/GqlFormattedValue"
        ytdChangePercent:
          $ref: "#/components/schemas/GqlFormattedValue"
    GqlSectorResearchReport:
      type: object
      properties:
        headline:
          type: string
        id:
          type: string
        investmentRating:
          type: string
        provider:
          type: string
        reportDate:
          type: string
        reportTitle:
          type: string
        reportType:
          type: string
        targetPrice:
          type: number
          format: double
        targetPriceStatus:
          type: string
      required:
        - id
...
    Industry:
      description: Typed industry identifier for the industry endpoint and custom screener queries.
      type: string
      enum:
        - agricultural-inputs
        - aluminum
        - coal
        - copper
        - farm-products
        - forest-products
        - gold
        - lumber-wood-production
        - other-industrial-metals-mining
        - other-precious-metals-mining
        - silver
        - steel
        - thermal-coal
        - uranium
        - apparel-manufacturing
        - apparel-retail
        - auto-truck-dealerships
        - auto-manufacturers
        - auto-parts
        - beverages-brewers
        - beverages-non-alcoholic
        - beverages-wineries-distilleries
        - confectioners
        - department-stores
        - discount-stores
        - electronic-gaming-multimedia
        - food-distribution
        - footwear-accessories
        - furnishings-fixtures-appliances
        - gambling
        - grocery-stores
        - home-improvement-retail
        - household-personal-products
        - internet-retail
        - leisure
        - lodging
        - luxury-goods
        - packaged-foods
        - personal-services
        - residential-construction
        - resorts-casinos
        - restaurants
        - specialty-retail
        - textile-manufacturing
        - tobacco
        - travel-services
        - oil-gas-drilling
        - oil-gas-ep
        - oil-gas-equipment-services
        - oil-gas-integrated
        - oil-gas-midstream
        - oil-gas-refining-marketing
        - solar
        - asset-management
        - banks-diversified
        - banks-regional
        - capital-markets
        - credit-services
        - financial-data-stock-exchanges
        - insurance-brokers
        - insurance-diversified
        - insurance-life
        - insurance-property-casualty
        - insurance-reinsurance
        - insurance-specialty
        - mortgage-finance
        - shell-companies
        - biotechnology
        - diagnostics-research
        - drug-manufacturers-general
        - drug-manufacturers-specialty-generic
        - health-information-services
        - healthcare-plans
        - medical-care-facilities
        - medical-devices
        - medical-distribution
        - medical-instruments-supplies
        - pharmaceutical-retailers
        - aerospace-defense
        - building-materials
        - building-products-equipment
        - business-equipment-supplies
        - chemical-manufacturing
        - chemicals
        - conglomerates
        - consulting-services
        - electrical-equipment-parts
        - engineering-construction
        - farm-heavy-construction-machinery
        - industrial-distribution
        - infrastructure-operations
        - integrated-freight-logistics
        - manufacturing-diversified
        - marine-ports-services
        - marine-shipping
        - metal-fabrication
        - paper-paper-products
        - pollution-treatment-controls
        - railroads
        - rental-leasing-services
        - security-protection-services
        - specialty-business-services
        - specialty-chemicals
        - specialty-industrial-machinery
        - staffing-employment-services
        - tools-accessories
        - trucking
        - waste-management
        - real-estate-development
        - real-estate-diversified
        - real-estate-services
        - reit-diversified
        - reit-healthcare-facilities
        - reit-hotel-motel
        - reit-industrial
        - reit-mortgage
        - reit-office
        - reit-residential
        - reit-retail
        - reit-specialty
        - communication-equipment
        - computer-hardware
        - consumer-electronics
        - data-analytics
        - electronic-components
        - electronics-computer-distribution
        - hardware-software-distribution
        - information-technology-services
        - internet-content-information
        - scientific-technical-instruments
        - semiconductor-equipment-materials
        - semiconductors
        - software-application
        - software-infrastructure
        - broadcasting
        - entertainment
        - publishing
        - telecom-services
        - utilities-diversified
        - utilities-independent-power-producers
        - utilities-regulated-electric
        - utilities-regulated-gas
        - utilities-regulated-water
        - utilities-renewable
        - closed-end-fund-debt
        - closed-end-fund-equity
        - closed-end-fund-foreign
        - exchange-traded-fund
...
    RateLimitError:
      description: "Rate limit error body, returned with HTTP 429 and a `Retry-After` header."
      type: object
      properties:
        error:
          description: "Always \"Rate limit exceeded\"."
          type: string
        message:
          description: Human-readable message including the retry delay.
          type: string
        retryAfterSeconds:
          description: Seconds until the next request would be admitted.
          type: integer
          format: int64
          minimum: 0
      required:
        - error
        - message
        - retryAfterSeconds
...
    Sector:
      description: Market sector types available on Yahoo Finance
      type: string
      enum:
        - technology
        - financial-services
        - consumer-cyclical
        - communication-services
        - healthcare
        - industrials
        - consumer-defensive
        - energy
        - basic-materials
        - real-estate
        - utilities
...
    ValueFormat:
      description: Value format for API responses
      type: string
      enum:
        - raw
        - pretty
        - both
...